+++ dimas/src/com/observation/mod.rs
@@ -1,0 +1,20 @@
+// Copyright © 2024 Stephan Kunz
+
+//! Module handles observation.
+//!
+
+// region:    	--- modules
+mod observable;
+mod observable_builder;
+mod observer;
+mod observer_builder;
+
+// flatten
+pub use observable::*;
+pub use observable_builder::*;
+pub use observer::*;
+pub use observer_builder::*;
+// endregion: 	--- modules
+
+// region:    	--- types
+// endregion: 	--- types