+++ dimas/src/com/pubsub/mod.rs
@@ -1,0 +1,20 @@
+// Copyright © 2024 Stephan Kunz
+
+//! Module handles publish/subscribe.
+//!
+
+// region:    	--- modules
+mod publisher;
+mod publisher_builder;
+mod subscriber;
+mod subscriber_builder;
+
+// flatten
+pub use publisher::*;
+pub use publisher_builder::*;
+pub use subscriber::*;
+pub use subscriber_builder::*;
+// endregion: 	--- modules
+
+// region:    	--- types
+// endregion: 	--- types